+++ src/react/atlascode/common/PrepareCommitTip.tsx
@@ -12,7 +12,9 @@ export const PrepareCommitTip: React.FC<PrepareCommitTipProps> = memo(({ classNa
                 Tip:{' '}
             </Box>{' '}
             You can have issue keys auto-added to your commit messages using{' '}
-            <Link href="https://bitbucket.org/snippets/atlassian/qedp7d">our prepare-commit-msg hook</Link>
+            <Link href="https://github.com/atlassian/atlascode/blob/main/scripts/prepare-commit-jira.sh">
+                our prepare-commit-msg hook
+            </Link>
         </Typography>
     );
 });